+ Moved Binding-related classes to Bindings folder
