+++ OSCADSharp/OSCADSharp/Bindings/Binding.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using OSCADSharp.Scripting;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace OSCADSharp.Scripting
+namespace OSCADSharp.Bindings
 {
     internal class Binding
     {
